Refactored scroll functions
